ENH: Creation of abstract classes for the python script
