+++ Modules/Scripted/ShapeAnalysisModule/CMakeLists.txt
@@ -12,6 +12,10 @@ set(MODULE_PYTHON_SCRIPTS
 set(MODULE_PYTHON_RESOURCES
   Resources/Icons/${MODULE_NAME}.png
   Resources/UI/${MODULE_NAME}.ui
+  CommonUtilities/__init__.py
+  CommonUtilities/utility.py
+  CommonUtilities/LogicMixin.py
+  CommonUtilities/PipelineMixin.py
   )
 
 #-----------------------------------------------------------------------------
