+++ Modules/Scripted/ShapeAnalysisModule/CommonUtilities/__init__.py
@@ -1,0 +1,3 @@
+from utility import *
+from LogicMixin import *
+from PipelineMixin import *
